requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
numpy>=1.21.0
opencv-python-headless>=4.5.3
scipy>=1.7.0
Pillow>=8.3.1